requirements.txt
openai-whisper
pandas
nltk
textblob
matplotlib
tqdm
gensim
pyLDAvis
